% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{cand}
\alias{cand}
\title{Base candidatos}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 181 rows and 17 columns.
}
\usage{
cand
}
\description{
Base candidatos
}
\keyword{datasets}
