% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_tidy}
\alias{da_tidy}
\title{Base tidy de conteúdo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 480956 rows and 27 columns.
}
\usage{
da_tidy
}
\description{
Base tidy de conteúdo
}
\keyword{datasets}
