% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\docType{data}
\name{da_xingamento}
\alias{da_xingamento}
\title{Base tidy de conteúdo ofensivo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 42222 rows and 28 columns.
}
\usage{
da_xingamento
}
\description{
Base tidy de conteúdo ofensivo
}
\keyword{datasets}
