% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{data_extracao}
\alias{data_extracao}
\title{Data extração}
\usage{
data_extracao(csv, rede = "tweets")
}
\arguments{
\item{csv}{path do arquivo de Twitter}

\item{rede}{c("tweets", "instagram")}
}
\value{
date
}
\description{
Extrai data a partir de path do arquivo
}
