% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_instagram}
\alias{get_instagram}
\title{Get instagram - diretório}
\usage{
get_instagram(arq)
}
\arguments{
\item{arq}{diretório com arquivos relacionados a Instagram}
}
\value{
tibble
}
\description{
Tidy dados de instagram
}
