% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_instagram2_}
\alias{get_instagram2_}
\title{Get instagram}
\usage{
get_instagram2_(csv)
}
\arguments{
\item{csv}{arquivo com dados de Instagram separados por ;}
}
\value{
tibble
}
\description{
Tidy dados de instagram
}
