% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_tweets}
\alias{get_tweets}
\title{Get tweets - diretório}
\usage{
get_tweets(arq)
}
\arguments{
\item{arq}{diretório com arquivos relacionados a Twitter}
}
\value{
tibble
}
\description{
Tidy diretório de csvs de tweets
}
