% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_youtube}
\alias{get_youtube}
\title{Get youtube - diretório}
\usage{
get_youtube(arq, type = "post")
}
\arguments{
\item{arq}{diretório com arquivos ods de YouTube}
}
\value{
tibble
}
\description{
Tidy dados de youtube
}
