% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_yt_channel}
\alias{get_yt_channel}
\title{Get youtube channel name}
\usage{
get_yt_channel(link)
}
\arguments{
\item{link}{id do canal de YouTube}
}
\value{
nome do canal
}
\description{
Tidy dados de youtube
}
