% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.R
\name{get_yt_comments}
\alias{get_yt_comments}
\title{Get youtube comments}
\usage{
get_yt_comments(yt_path, id)
}
\arguments{
\item{yt_path}{caminho para arquivo ods com dados de YouTube}

\item{id}{id do vídeo}
}
\value{
tibble
}
\description{
Tidy dados de youtube
}
